backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
gunicorn>=21.2.0
python-multipart>=0.0.7
pydantic>=2.6.0
vaderSentiment>=3.3.2
pandas>=2.2.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
bleach>=6.1.0
fastapi-limiter>=0.1.5
aioredis>=2.0.1
redis>=5.0.1
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
psutil>=5.9.0
python-json-logger>=2.0.7
types-python-jose>=3.3.4
types-passlib>=1.7.7
types-redis>=4.6.0